requirements.txt
neo4j
python-dotenv
sentence-transformers
requests
flask
pymupdf